Move role and role config group result parsing to separate modules. Update general output parsing function.
