+++ plugins/module_utils/role_config_group_utils.py
@@ -12,35 +12,24 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""
-A common functions for Cloudera Manager cluster management
-"""
-
 from ansible_collections.cloudera.cluster.plugins.module_utils.cm_utils import (
     normalize_output,
 )
 
-from cm_client import ApiCluster
+from cm_client import ApiRoleConfigGroup
 
 
-CLUSTER_OUTPUT = [
+ROLE_CONFIG_GROUP = [
     "name",
+    "role_type",
+    "base",
     "display_name",
-    # "full_version",
-    "maintenance_mode",
-    "maintenance_owners",
-    # "services",
-    # "parcels",
-    "entity_status",
-    "uuid",
-    # "data_context_refs",
-    "cluster_type",
-    "tags",
+    # "service_ref",
 ]
 
 
-def parse_cluster_result(cluster: ApiCluster) -> dict:
-    # Retrieve full_version as version
-    output = dict(version=cluster.full_version)
-    output.update(normalize_output(cluster.to_dict(), CLUSTER_OUTPUT))
+def parse_role_config_group_result(role_config_group: ApiRoleConfigGroup) -> dict:
+    # Retrieve only the service identifier
+    output = dict(service_name=role_config_group.service_ref.service_name)
+    output.update(normalize_output(role_config_group.to_dict(), ROLE_CONFIG_GROUP))
     return output
