+++ plugins/module_utils/role_utils.py
@@ -12,35 +12,37 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""
-A common functions for Cloudera Manager cluster management
-"""
-
 from ansible_collections.cloudera.cluster.plugins.module_utils.cm_utils import (
     normalize_output,
 )
 
-from cm_client import ApiCluster
+from cm_client import ApiRole
 
-
-CLUSTER_OUTPUT = [
-    "name",
-    "display_name",
-    # "full_version",
+ROLE_OUTPUT = [
+    "commission_state",
+    "config_staleness_status",
+    "ha_status",
+    "health_checks",
+    "health_summary",
+    # "host_ref",
     "maintenance_mode",
     "maintenance_owners",
-    # "services",
-    # "parcels",
-    "entity_status",
-    "uuid",
-    # "data_context_refs",
-    "cluster_type",
+    "name",
+    # "role_config_group_ref",
+    "role_state",
+    # "service_ref",
     "tags",
+    "type",
+    "zoo_keeper_server_mode",
 ]
 
 
-def parse_cluster_result(cluster: ApiCluster) -> dict:
-    # Retrieve full_version as version
-    output = dict(version=cluster.full_version)
-    output.update(normalize_output(cluster.to_dict(), CLUSTER_OUTPUT))
+def parse_role_result(role: ApiRole) -> dict:
+    # Retrieve only the host_id, role_config_group, and service identifiers
+    output = dict(
+        host_id=role.host_ref.host_id,
+        role_config_group_name=role.role_config_group_ref.role_config_group_name,
+        service_name=role.service_ref.service_name,
+    )
+    output.update(normalize_output(role.to_dict(), ROLE_OUTPUT))
     return output
