+++ plugins/modules/service_role_config_group_config_info.py
@@ -16,7 +16,6 @@
 
 from ansible_collections.cloudera.cluster.plugins.module_utils.cm_utils import (
     ClouderaManagerModule,
-    parse_role_config_group_result,
 )
 
 from cm_client import (
@@ -25,6 +24,7 @@ from cm_client import (
     ServicesResourceApi,
 )
 from cm_client.rest import ApiException
+
 
 ANSIBLE_METADATA = {
     "metadata_version": "1.1",
