+++ plugins/modules/service_role_config_group_info.py
@@ -16,6 +16,9 @@
 
 from ansible_collections.cloudera.cluster.plugins.module_utils.cm_utils import (
     ClouderaManagerModule,
+)
+
+from ansible_collections.cloudera.cluster.plugins.module_utils.role_config_group_utils import (
     parse_role_config_group_result,
 )
 
@@ -25,6 +28,7 @@ from cm_client import (
     ServicesResourceApi,
 )
 from cm_client.rest import ApiException
+
 
 ANSIBLE_METADATA = {
     "metadata_version": "1.1",
