+++ plugins/modules/service_role_info.py
@@ -18,11 +18,15 @@ import json
 
 from ansible_collections.cloudera.cluster.plugins.module_utils.cm_utils import (
     ClouderaManagerModule,
+)
+
+from ansible_collections.cloudera.cluster.plugins.module_utils.role_utils import (
     parse_role_result,
 )
 
 from cm_client import ClustersResourceApi, RolesResourceApi, ServicesResourceApi
 from cm_client.rest import ApiException
+
 
 ANSIBLE_METADATA = {
     "metadata_version": "1.1",
